implemented end stop polling.
This is needed for boards that connected the end stops to cpu pins that
can not create an interrupt on a logic level change of the pin. Dear
Board designes: Please avoid to use this! Map the end switched to a pin
with pin change interrupt. That will give you much better perfromance!